Fix build error
